requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
bcrypt==4.1.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
orjson==3.9.15
packaging==23.2
parso==0.8.3
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
bcrypt==4.1.2
black==24.2.0
botocore==1.34.61
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
orjson==3.9.15
packaging==23.2
parso==0.8.3
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
